# build: make msd flash faster