This is mainly for cases where msd flash is used instead of openocd.
The basic idea is to run the `touchSerialPortAt1200bps()` in parallel with the build.

It makes the environment a little more comfortable for tinygopher, which doesn't have a debugger.
If the build takes more than 3 seconds, it will be at least 3 seconds faster.
If the build takes less than 3 seconds, there is not much improvement.


* Core i5-5200 + ubuntu 20.04
    * before : 8.6 sec
    * after : 3.5 sec
* Ryzen 7 3700X + Windows10
    * before : 5.0 sec
    * after : 4.5 sec